refactor(runtime): split object/field_set_by_name.rs under the 2000-line cap (#7402)
Pure mechanical relocation of `js_object_set_field_by_name` and its
neighbours into topical sub-modules. The file sat at exactly 2000 of 2000
allowed lines, so the next line added anywhere in it turned the required
`lint` context red.

No behaviour change: every moved statement is byte-identical, and the
split point for the tail is the `RuntimeHandleScope` creation, so the
`refresh_roots_after_alloc!()` macro (#7341) and all 16 of its call
sites — plus both `GC_STORE_AUDIT(INIT)` markers — stay together in one
file with their ordering untouched.
